Change schema of package voucher. TEST ALL /packages endpoint. TEST POST /orders and /payment

+++ src/resources/package/package.service.ts
@@ -1,10 +1,14 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { PackageVoucherRepository } from 'src/infrastructure/persistence/package/package.repository';
-import { CreatePackageVoucherDto } from './dto/create-package.dto';
+import {
+  CreatePackageVoucherDto,
+  PACKAGE_FILE_FIELD,
+} from './dto/create-package.dto';
 import {
   PackageImgCreateInput,
   PackageImgDomain,
   PackageRewardVoucherCreateInput,
+  PackageRewardVoucherDomain,
   PackageVoucherCreateInput,
   PackageVoucherDomain,
 } from './domain/package-voucher.domain';
@@ -19,9 +23,23 @@ import {
   packageVoucherTermAndCondENCreateInput,
   packageVoucherTermAndCondTHCreateInput,
 } from './domain/package-voucher-term-cond.domain';
-import { UpdatePackageVoucherDto } from './dto/update-package.dto';
+import {
+  UpdatePackageRewardVoucherDto,
+  UpdatePackageVoucherDto,
+} from './dto/update-package.dto';
 import { PackageImgRepository } from 'src/infrastructure/persistence/package/package-img.repository';
-import { ProductDomainHelper } from '@resources/account/dto/product.helper';
+import { ProductDomainHelper } from 'src/common/product.helper';
+import {
+  VoucherCategoryDomain,
+  VoucherDomain,
+} from '@resources/voucher/domain/voucher.domain';
+import { ObjectHelper } from '@utils/services/object.helper';
+import { CalculatorService } from '@utils/services/calculator.service';
+import {
+  PackageSellDateQueryEnum,
+  PackageStatusQueryEnum,
+} from './dto/get-package.dto';
+import { EnumCheckerHelper } from '@utils/services/enum-checker.helper';
 
 @Injectable()
 export class PackageVoucherService {
@@ -40,13 +58,39 @@ export class PackageVoucherService {
     packageImg,
   }: {
     data: CreatePackageVoucherDto;
-    mainImg: Express.Multer.File;
+    mainImg: Express.Multer.File[];
     packageImg?: Express.Multer.File[];
   }): Promise<PackageVoucherDomain> {
-    const idList = [
-      data.quotaVoucherId,
-      ...data.rewardVoucherId.filter((item) => item !== data.quotaVoucherId),
-    ];
+    if (!mainImg || mainImg.length === 0)
+      throw ErrorApiResponse.badRequest('Main image for voucher is required.');
+    // Extract reward voucher ID from package voucher data
+    const { rewardVouchers, termAndCondTh, termAndCondEn, ...restData } = data;
+
+    if (termAndCondTh.length === 0 && termAndCondEn.length === 0)
+      throw ErrorApiResponse.badRequest(
+        `Please provide value for term and condition field.`,
+      );
+
+    const rewardVoucherIdSet = new Set<VoucherDomain['id']>();
+    const rewardVoucherData: PackageRewardVoucherCreateInput[] = [];
+    // Extract reward voucher ID from package voucher data
+    const packageId = String(this.uuidService.make());
+    const idList = [data.quotaVoucherId];
+    for (const item of data.rewardVouchers) {
+      if (rewardVoucherIdSet.has(item.voucherId))
+        throw ErrorApiResponse.badRequest(
+          `Reward voucher ID: ${item.voucherId} was duplicated. If desired to add more amount of the same voucher, please add the number in the amount of the desired voucher ID field.`,
+        );
+      rewardVoucherIdSet.add(item.voucherId);
+      rewardVoucherData.push({
+        id: String(this.uuidService.make()),
+        amount: item.amount,
+        packageId,
+        rewardVoucherId: item.voucherId,
+      });
+      if (item.voucherId !== data.quotaVoucherId) idList.push(item.voucherId);
+    }
+
     // Check first if the voucher ID provided as
     // quota and reward is existing.
     const isAllVouchersExist =
@@ -69,7 +113,7 @@ export class PackageVoucherService {
     // to store in database.
     const allImgBuffer: Express.Multer.File[] = [];
 
-    if (mainImg) allImgBuffer.push(mainImg);
+    allImgBuffer.push(mainImg[0]);
     if (packageImg && packageImg.length > 0) allImgBuffer.push(...packageImg);
 
     if (allImgBuffer.length < 1) throw ErrorApiResponse.conflictRequest();
@@ -88,21 +132,11 @@ export class PackageVoucherService {
     // ------- SECOND PART : PREPARE INFORMATION -------
 
     // Extract reward voucher ID from package voucher data
-    const { rewardVoucherId, termAndCondTh, termAndCondEn, ...restData } = data;
-
+    // Extract reward voucher ID from package voucher data
     const packageData: PackageVoucherCreateInput = {
-      id: String(this.uuidService.make()),
+      id: packageId,
       ...restData,
     };
-
-    const rewardVoucherData: PackageRewardVoucherCreateInput[] =
-      rewardVoucherId.map((item) => {
-        return {
-          id: String(this.uuidService.make()),
-          rewardVoucherId: item,
-          packageId: packageData.id,
-        };
-      });
 
     const packageImgCreateData: PackageImgCreateInput[] =
       allPackageImgLinks.map((item, index) => {
@@ -142,15 +176,50 @@ export class PackageVoucherService {
   }
 
   async getAllPackageVoucher({
+    category,
     cursor,
+    status,
+    sellDate,
   }: {
     cursor?: PackageVoucherDomain['id'];
+    category?: VoucherCategoryDomain['name'];
+    status?: PackageStatusQueryEnum;
+    sellDate?: PackageSellDateQueryEnum;
   }): Promise<PackageVoucherDomain[]> {
+    const statusToQuery = this.checkStatusQuery(status);
+    const sellDateQuery = this.checkSellDateQuery(sellDate);
+
     if (cursor && !isUUID(cursor))
       throw ErrorApiResponse.badRequest(
         `Cursor: ${cursor} is not valid data type for searching.`,
       );
-    return this.packageVoucherRepository.findManyPackageVoucher({ cursor });
+
+    if (status) {
+    }
+    return this.packageVoucherRepository.findManyPackageVoucher({
+      cursor,
+      category,
+      status: statusToQuery,
+      sellDate: sellDateQuery,
+    });
+  }
+
+  private checkStatusQuery(status: PackageStatusQueryEnum) {
+    return EnumCheckerHelper.getEnumValueOrThrow(
+      PackageStatusQueryEnum,
+      status,
+      PackageStatusQueryEnum.ACTIVE,
+    );
+  }
+
+  private checkSellDateQuery(
+    sellDate: PackageSellDateQueryEnum,
+  ): PackageSellDateQueryEnum {
+    return EnumCheckerHelper.getEnumValueOrThrow(
+      PackageSellDateQueryEnum,
+      sellDate,
+      PackageSellDateQueryEnum.NOW,
+    );
   }
 
   async getPackageVoucherById(
@@ -160,7 +229,13 @@ export class PackageVoucherService {
       throw ErrorApiResponse.badRequest(
         `${packageId} is not valid data type for searching.`,
       );
-    return this.packageVoucherRepository.findPackageVoucherById(packageId);
+    const packageVoucher =
+      await this.packageVoucherRepository.findPackageVoucherById(packageId);
+    if (!packageVoucher)
+      throw ErrorApiResponse.notFoundRequest(
+        `Package ID: ${packageId} could not be found.`,
+      );
+    return packageVoucher;
   }
 
   async updatePackageVoucher(
@@ -178,9 +253,133 @@ export class PackageVoucherService {
         `Package ID: ${data.id} could not be found.`,
       );
 
+    // Check the update date data if provided
     ProductDomainHelper.checkUsableAndSellTime(data, isPackageExist, 'package');
 
+    // Check the other update data if provided
+    ProductDomainHelper.checkUpdateData(data, isPackageExist, 'package');
+    if (!ObjectHelper.isObjectEmpty(data.rewardVouchers)) {
+      await this.checkRewardVoucherBeforeUpdate(
+        data.rewardVouchers,
+        isPackageExist.rewardVouchers,
+      );
+    }
+
     return this.packageVoucherRepository.updatePackageVoucher(data);
+  }
+
+  private async checkRewardVoucherBeforeUpdate(
+    data: UpdatePackageRewardVoucherDto,
+    existRewardVouchers: PackageVoucherDomain['rewardVouchers'],
+  ) {
+    const { addRewardVouchers, removedRewardIds, update } = data;
+
+    const rewardVoucherIdMap = new Map<
+      PackageRewardVoucherDomain['id'],
+      number
+    >();
+    existRewardVouchers.forEach((item) => {
+      rewardVoucherIdMap.set(item.id, item.amount);
+    });
+
+    if (addRewardVouchers && addRewardVouchers.length === 0)
+      throw ErrorApiResponse.badRequest(
+        `Field addRewardVouchers does not have any provide value.`,
+      );
+
+    if (update && update.length === 0)
+      throw ErrorApiResponse.badRequest(
+        `Field update does not have any provide value.`,
+      );
+
+    if (removedRewardIds && removedRewardIds.length === 0)
+      throw ErrorApiResponse.badRequest(
+        `Field removedVoucherIds does not have any provide value.`,
+      );
+
+    // ADDING NEW REWARD VOUCER PATH
+    if (addRewardVouchers && addRewardVouchers.length > 0) {
+      const voucherMap = new Map<VoucherDomain['id'], boolean>();
+      const allRewardVoucherIdList = [];
+      addRewardVouchers.forEach((item) => {
+        if (voucherMap.has(item.voucherId))
+          throw ErrorApiResponse.badRequest(
+            `Reward voucher ID: ${item.voucherId} in adding request is duplicated.`,
+          );
+
+        if (rewardVoucherIdMap.has(item.voucherId))
+          throw ErrorApiResponse.badRequest(
+            `Reward voucher ID: ${item.voucherId} is already part of the package. Please check if the reward voucher ID is correct or if request desire to update, please update instead.`,
+          );
+
+        voucherMap.set(item.voucherId, true);
+        allRewardVoucherIdList.push(item.voucherId);
+      });
+      const isAllVoucherExist = await this.voucherService.getVoucherByIds(
+        allRewardVoucherIdList,
+      );
+
+      if (isAllVoucherExist.length !== allRewardVoucherIdList.length) {
+        const notFoundId = isAllVoucherExist.filter((itemId) =>
+          allRewardVoucherIdList.filter((ele) => ele.id !== itemId),
+        );
+        throw ErrorApiResponse.conflictRequest(
+          `Some reward voucher ID: ${notFoundId.join(', ')} could not be found.`,
+        );
+      }
+    }
+
+    if (update && update.length > 0) {
+      const updateIdMap = new Map<VoucherDomain['id'], boolean>();
+      update.forEach((item) => {
+        const existAmount = rewardVoucherIdMap.get(item.rewardId);
+        if (!existAmount)
+          throw ErrorApiResponse.conflictRequest(
+            `Reward voucher ID: ${item.rewardId} is not part of the package. Please check if the reward voucher ID is correct or it could not be found.`,
+          );
+
+        if (updateIdMap.has(item.rewardId))
+          throw ErrorApiResponse.badRequest(
+            `Reward voucher ID: ${item.rewardId} in update request is duplicated.`,
+          );
+
+        if (item.amount === existAmount)
+          throw ErrorApiResponse.badRequest(
+            `Update amount is the same as the existing amount in reward ID: ${item.rewardId}. `,
+          );
+
+        if (item.amount <= 0)
+          throw ErrorApiResponse.conflictRequest(
+            `The new update amount will cause the reward voucher amount to be 0. If desired request is to delete. Please make the delete request instead.`,
+          );
+        updateIdMap.set(item.rewardId, true);
+      });
+    }
+
+    // REMOVE REWARD VOUCHERS
+    if (removedRewardIds && removedRewardIds.length > 0) {
+      if (removedRewardIds.length >= existRewardVouchers.length) {
+        throw ErrorApiResponse.conflictRequest(
+          `Package voucher must have at least one reward voucher. Desired delete amount: ${removedRewardIds.length}, Current reward voucher amount: ${existRewardVouchers.length}`,
+        );
+      }
+      const removeVoucherIdMap = new Map<VoucherDomain['id'], boolean>();
+      removedRewardIds.forEach((item) => {
+        if (!rewardVoucherIdMap.has(item))
+          throw ErrorApiResponse.conflictRequest(
+            `Reward voucher ID: ${item} is not part of the package. Please check if the reward voucher ID is correct or it could not be found.`,
+          );
+
+        if (removeVoucherIdMap.has(item))
+          throw ErrorApiResponse.conflictRequest(
+            `Reward voucher ID: ${item} in remove request is duplicated.`,
+          );
+
+        removeVoucherIdMap.set(item, true);
+      });
+    }
+
+    return;
   }
 
   async deletePackageVoucherById(
@@ -189,6 +388,19 @@ export class PackageVoucherService {
     if (!packageId || !isUUID(packageId, 7))
       throw ErrorApiResponse.conflictRequest(
         `${packageId} is not the valid type of data for this request.`,
+      );
+
+    const isPackageExist =
+      await this.packageVoucherRepository.findPackageVoucherById(packageId);
+
+    if (!isPackageExist)
+      throw ErrorApiResponse.notFoundRequest(
+        `Package ID: ${packageId} could not be found.`,
+      );
+
+    if (isPackageExist.deletedAt)
+      throw ErrorApiResponse.conflictRequest(
+        `Package ID: ${packageId} is already deleted.`,
       );
 
     return this.packageVoucherRepository.deletePackageVoucherById(packageId);
@@ -202,11 +414,21 @@ export class PackageVoucherService {
     id: PackageVoucherDomain['id'],
     files: Express.Multer.File[],
   ): Promise<PackageImgDomain[]> {
+    if (!files || files.length === 0)
+      throw ErrorApiResponse.badRequest(
+        `${PACKAGE_FILE_FIELD.PACKAGE_IMG} field is required for this request.`,
+      );
+
     const isPackageExist =
       await this.packageVoucherRepository.findPackageVoucherById(id);
     if (!isPackageExist)
       throw ErrorApiResponse.notFoundRequest(
         `Package ID: ${id} could not be found.`,
+      );
+
+    if (isPackageExist.deletedAt)
+      throw ErrorApiResponse.conflictRequest(
+        `Package ID: ${id} is already deleted.`,
       );
 
     const uploadedImgPath = await Promise.all(
@@ -249,6 +471,8 @@ export class PackageVoucherService {
       filePath: file.path,
       bucketDir: s3BucketDirectory.packageImg,
     });
+
+    await this.mediaService.deleteFile(isPackageImgExist.imgPath);
     return this.packageImgRepository.update(id, uploadedImgPath);
   }
